move ReleaseInfo to utils
